backend/requirements.txt
fastapi 
uvicorn 
supabase 
python-dotenv 
bcrypt 
pydantic 
